+++ tests/testthat/test-consistency-selection.R
@@ -1,4 +1,4 @@
-context("5. Test consistency of function selection()")
+context("6. Test consistency of function selection()")
 
 # jfa version 0.1.0