+++ tests/testthat/test-consistency-workflow.R
@@ -1,4 +1,4 @@
-context("6. Test consistency of workflow functionality")
+context("7. Test consistency of workflow functionality")
 
 # jfa version 0.1.0