+++ tests/testthat/test-validation-touwhoogduin.R
@@ -1,4 +1,4 @@
-context("12. Benchmark against Touw and Hoogduin (2011)")
+context("13. Benchmark against Touw and Hoogduin (2011)")
 
 # Touw, P., and Hoogduin, L. (2011). Statistiek voor audit en controlling. Boom uitgevers, Amsterdam.